% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count}
\alias{cell_count}
\title{Count filled cells}
\usage{
cell_count(data, xmin, xmax, ymin, ymax, zmin, zmax, ngrid)
}
\arguments{
\item{ngrid}{}
}
\value{

}
\description{
Count filled cells
}
