% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count_1d}
\alias{cell_count_1d}
\title{Count filled cells in 1D}
\usage{
cell_count_1d(pts, xmin, xmax, n)
}
\arguments{
\item{pts}{data}

\item{xmin}{minimum x-value}

\item{xmax}{maximum x-value}

\item{n}{multiplier}
}
\value{
Number of filled cells
}
\description{
Count filled cells in 1D
}
