% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count_2d}
\alias{cell_count_2d}
\title{Count filled cells in 2D}
\usage{
cell_count_2d(pts, xmin, xmax, ymin, ymax, n)
}
\arguments{
\item{pts}{data}

\item{xmin}{minimum x-value}

\item{xmax}{maximum x-value}

\item{ymin}{minimum y-value}

\item{ymax}{maximum y-value}

\item{n}{multiplier}
}
\value{
Number of filled cells
}
\description{
Count filled cells in 2D
}
