% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Centroid}
\usage{
centroid(XYZCoords)
}
\arguments{
\item{XYZCoords}{A data frame with x, y, and z coordinates}
}
\value{
the coordinates of the centroid
}
\description{
Calculates the centroid for a given set of XYZ coordinates.
}
\examples{
data <- mesh_to_points(mcap)
centroid(data)
}
