% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexity.R
\name{convexity}
\alias{convexity}
\title{Convexity}
\usage{
convexity(mesh)
}
\arguments{
\item{mesh}{A mesh3d object.}
}
\value{
The convexity value
}
\description{
Calculates the convexity of a 3D mesh.
}
\details{
The ratio of the volume of the object and the volume of the convex
hull around the object. Objects with fewer concavities will be closer to 1.
}
\examples{
convexity(mcap)
}
