% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_tilt.R
\name{correct_tilt}
\alias{correct_tilt}
\title{Correct the tilt of the DEM}
\usage{
correct_tilt(data)
}
\arguments{
\item{data}{digital elevation model (DEM)}
}
\value{
DEM with corrected tilt
}
\description{
Correct the tilt of the DEM
}
