% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_dem.R
\name{crop_dem}
\alias{crop_dem}
\title{Crop DEM around Points}
\usage{
crop_dem(data, x0, y0, L, plot = FALSE)
}
\arguments{
\item{data}{a DEM in RasterLayer format}

\item{x0}{A value or vector of center x coordinate(s)}

\item{y0}{A value or vector of center y coordinate(s)}

\item{L}{Size of square to cut out of DEM}

\item{plot}{Plot the DEM and the cropped sections (TRUE or FALSE)}
}
\value{
A cropped RasterLayer or a list of RasterLayers
}
\description{
Crop DEM around Points
}
\examples{
# around one point
dem_cropped <- crop_dem(horseshoe, -468, 1266, L = 2)
raster::plot(dem_cropped)

# around multiple points
points <- data.frame(x = c(-467, -465, -466), y = c(1270, 1265, 1268))
dem_list <- crop_dem(horseshoe, points$x, points$y, L = 1, plot = TRUE)
}
