% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf.R
\name{csf}
\alias{csf}
\title{Calculates mechanical vulnerability of cantilever type structure to flow}
\usage{
csf(mesh, z_min, res)
}
\arguments{
\item{mesh}{mesh object}

\item{z_min}{(optional) set the z plane about which csf should be calculated}

\item{res}{(optional) set the resolution}
}
\value{
A list containing the colony shape factor, the parallel (d1) and perpendicular (d2) diameters of the attachment point, and the second moment of area relative to flow.
}
\description{
Calculates mechanical vulnerability of cantilever type structure to flow
}
\details{
This function calculates the mechanical vulnerability of a structural
element, like a hard coral colony, to fluid flow. While developed for corals, and originally
called the Colony Shape Factor (CSF), the function is applicable to any attached,
rigid cantilever type structure. CSF is dimensionless and can be used to compare the
vulnerability among structures. Mechanistically, if the CSF of a structure becomes greater
than the dislodgement mechanical threshold, breakage occurs. This threshold is a function of
material tensile strength and inversely related to fluid velocity and density.
}
\examples{
csf(mcap, z_min=-3.65)
}
\references{
Madin JS and Connolly SR (2006) Ecological consequences of major hydrodynamic disturbances on coral reefs. Nature. 444:477-480.
}
