% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_points.R
\name{dem_to_points}
\alias{dem_to_points}
\title{Transform dem to 3D pointcloud of raster corners.}
\usage{
dem_to_points(dem, bh = NULL, parallel = F)
}
\arguments{
\item{dem}{A digital elevation model in raster format}
}
\value{
A 3D point cloud for raster cell corners
}
\description{
Transform dem to 3D pointcloud of raster corners.
}
