% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_drop.R
\name{detect_drop}
\alias{detect_drop}
\title{detect a sudden drop, edge, or overhang in a DEM}
\usage{
detect_drop(data, d = 0.1)
}
\arguments{
\item{data}{DEM of class RasterLayer}

\item{d}{threshold height difference to define a drop}
}
\value{
A RasterLayer marking edges. Values indicate maximum height difference of surrounding cells
}
\description{
detect a sudden drop, edge, or overhang in a DEM
}
\examples{
library(raster)
library(habtools)
dem <- detect_drop(horseshoe, d = 0.2)
plot(horseshoe)
plot(dem)

}
