% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\title{Calculate fractal dimension using cube counting}
\usage{
extent(data)
}
\arguments{
\item{data}{object of class mesh3d}
}
\value{
A value, the extent of the mesh
}
\description{
Calculate fractal dimension using cube counting
}
\details{
This function calculates the extent or largest length of the bounding box of a mesh
}
\examples{
extent(mcap)

}
