% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd}
\alias{fd}
\title{fractal dimension}
\usage{
fd(data, method, x, y, L, Lvec)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer or a triangular mesh of class mesh3d.}

\item{method}{If data is a RasterLayer methods "hvar" or "area" are allowed. Defaults to "hvar".}

\item{x}{Bottom-left of bounding box.}

\item{y}{Bottom-left of bounding box.}

\item{L}{Bounding box extent (i.e., side length).}

\item{Lvec}{scales to use for calculation}

\item{parallel}{Use parallel processing (Defaults to FALSE)}

\item{ncores}{number of cores to use if parallel = T.
(Defaults to umber of available cores - 1)}
}
\value{
A value for fractal dimension, hopefully between 2 and 3.
}
\description{
fractal dimension
}
\details{
Calculates fractal dimension using the height variation method.
\code{data} can be any \code{data.frame} with columns labeled \code{l} and \code{h} for
grid cell length and height range of that cell, respectively.
A rule of thumb is that \code{l} should range two orders of magnitude.
However, large ranges also
average-out fractal dimension of a surface that might have
phase transitions, and therefore a thorough exploration of height ranges is suggested using the \code{diagnostic} plot, which plots all the methods.
These methods will converge on purely fractal surfaces.
}
\examples{
fd(horseshoe, method = "hvar", x=-470, y=1266, L=2, Lvec = c(0.125, 0.25, 0.5, 1, 2))
fd(data)
fd(data, method = "ends")
fd_hvar(data, method = "median", plot = TRUE)

}
