% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_area.R
\name{fd_area}
\alias{fd_area}
\title{Calculate fractal dimension using the surface area method}
\usage{
fd_area(data, lvec = NULL, keep_data = FALSE, plot = FALSE, scale = FALSE)
}
\arguments{
\item{data}{DEM of class "RasterLayer" or mesh of class "mesh3d".}

\item{lvec}{Vector of scales to include in calculation.}

\item{keep_data}{Logical. Keep data? Default is FALSE.}

\item{plot}{Logical. Plot regression line and area data?  Defaults to FALSE.}

\item{scale}{Logical. Rescale height values to fit the extent? Only relevant for DEMs. Defaults to FALSE.}
}
\value{
Either a value or a list if keep_data = TRUE.
}
\description{
Calculate fractal dimension using the surface area method
}
\examples{
fd_area(mcap, lvec = c(0.01, 0.02, 0.04, 0.08, 0.16))
fd_area(horseshoe, lvec = c(0.1, 0.2, 0.4, 0.8, 1.6))

}
