% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_boxes.R
\name{fd_boxes}
\alias{fd_boxes}
\title{Calculate fractal dimension using box counting}
\usage{
fd_boxes(pts, lvec = NULL, plot = FALSE, keep_data = FALSE)
}
\arguments{
\item{lvec}{scales to use for calculation (i.e. box sizes)}

\item{plot}{Plot number of filled boxes at different scales. Default = FALSE.}

\item{keep_data}{Keep calculation data. Default = FALSE.}

\item{data}{data frame of xy coordinates}
}
\value{
A \code{data.frame} of number of boxes (\code{n}) intersecting xy coordinates at different box sizes (\code{l}) and a fractal dimension value. Note that box size is the length of a side.
}
\description{
Calculate fractal dimension using box counting
}
\details{
This function calculates fractal dimension using the cube counting method.
Based on lvec, cubes of different sizes are defined and the function counts mesh points that fall within each cube.
It is recommended that to specify the maximum value of \code{lvec} so that the largest box encapsulates the entire object.
The smallest scale included in \code{lvec} should not be smaller than the resolution of your object.
Most meshes are not perfectly fractal, and so use the \code{plot} parameter to look for scale transitions.
}
\examples{
mcap_2d <- mesh_to_2d(mcap)

fd_boxes(mcap_2d, lvec = c(0.05, 0.1, 0.25, 0.5))

}
