% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_diagnose.R
\name{fd_diagnose}
\alias{fd_diagnose}
\title{Investigates varying fractal dimensions across scale}
\usage{
fd_diagnose(data, lvec = NULL, method, ...)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer}

\item{lvec}{scales to use for calculation}

\item{method}{Possible methods are:"hvar", "area", and "sd".}

\item{...}{Additional arguments of see \code{\link[=fd]{fd()}}}
}
\value{
A list with fractal dimension across scales, mean fractal dimension, and sd of fractal dimensions across scales.
}
\description{
Investigates varying fractal dimensions across scale
}
\examples{
fd_diagnose(horseshoe, lvec = c(0.05, 0.1, 0.2, 0.4), method = "area")

}
