% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_sd.R
\name{fd_sd}
\alias{fd_sd}
\title{Calculate fractal Dimension using the standard deviation method}
\usage{
fd_sd(
  data,
  x,
  y,
  lvec,
  L,
  regmethod = "mean",
  keep_data = FALSE,
  plot = FALSE,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer}

\item{x}{Bottom-left of bounding box}

\item{y}{Bottom-left of bounding box}

\item{lvec}{scales to use for calculation}

\item{L}{Bounding box extent (i.e., side length)}

\item{regmethod}{Method to use for linear regression between scale (lvec) and height range. One of \code{raw} (all data), \code{mean} (default) \code{median} or \code{ends} (minimum and maximum scale only)}

\item{keep_data}{Keep the data used for fd calculation? defaults to FALSE}

\item{plot}{Diagnostic plot}
}
\value{
A value for fractal dimension, normally between 2 and 3.
}
\description{
Calculate fractal Dimension using the standard deviation method
}
\details{
Calculates fractal dimension using the standard deviation method,
an analogue of the variation method, but using the standard deviation in height per grid cell instead of the full height range.
A rule of thumb is that \code{l} should range at least an order of magnitude.
However, large ranges also
average-out fractal dimension of a surface that might have
phase transitions, and therefore a thorough exploration of height ranges is suggested using the \code{plot}.
\code{regmethod} specifies whether data is summarized by taking the mean or median of height ranges across scales or all data is used.
\code{regmethod} "raw" is not recommended because the regression will give much more weight to the lower scales that include more points and likely underestimate D.
}
\examples{
fd_sd(horseshoe, x=-470, y=1266, L=2, lvec = c(0.05, 0.1, 0.25, 0.5, 1, 2))

}
