% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\docType{data}
\name{horseshoe}
\alias{horseshoe}
\title{Horseshoe reef}
\format{
A 3999 by 3999 digital elevation model (raster file)
\describe{
\item{values}{water depth}
\item{resolution}{0.002 m}
...
}
}
\usage{
horseshoe
}
\description{
A data set containing the prices and other attributes of almost 54,000
diamonds.
}
\examples{
plot(horseshoe)
}
\keyword{datasets}
