% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcap.R
\docType{data}
\name{mcap}
\alias{mcap}
\title{Montipora capitata}
\format{
A 3999 by 3999 digital elevation model (raster file)
\describe{
\item{values}{water depth}
\item{resolution}{0.002 m}
...
}
}
\usage{
mcap
}
\description{
A data set containing the prices and other attributes of almost 54,000
diamonds.
}
\examples{
library(rgl)
plot3d(mcap)
}
\keyword{datasets}
