% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcap2.R
\docType{data}
\name{mcap2}
\alias{mcap2}
\title{Montipora capitata 2}
\format{
mesh3d object.
}
\usage{
mcap2
}
\description{
A remeshed version of mcap with resolution = 0.002.
}
\examples{
library(rgl)
plot3d(mcap2)
}
\keyword{datasets}
