% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_points.R
\name{mesh_to_points}
\alias{mesh_to_points}
\title{Transform DEM to 3D pointcloud of raster corners}
\usage{
mesh_to_points(mesh)
}
\arguments{
\item{mesh}{A digital elevation model in raster format}
}
\value{
A 3D point cloud
}
\description{
Transform DEM to 3D pointcloud of raster corners
}
