% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdh.R
\name{rdh}
\alias{rdh}
\title{Calculate rugosity, fractal dimension and height for a DEM}
\usage{
rdh(data, x, y, L, lvec, ...)
}
\arguments{
\item{data}{A dem of class RasterLayer}

\item{x}{minimum x-value}

\item{y}{minimum y-value}

\item{L}{Extent}

\item{lvec}{Scales to include in fd calculation see \code{\link[=fd]{fd()}}}

\item{...}{Additional arguments see \code{\link[=fd]{fd()}}}
}
\value{
a dataframe with the three complexity measures
}
\description{
Calculate rugosity, fractal dimension and height for a DEM
}
\details{
uses hvar method for rugosity and fractal dimension calculations.
}
\examples{
rdh(horseshoe, x = -470, y = 1266, L = 2, lvec = c(0.125, 0.25, 0.5, 1, 2))
}
\seealso{
\code{\link[=fd]{fd()}}

\code{\link[=rg]{rg()}}

\code{\link[=hr]{hr()}}
}
