% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdh_theory.R
\name{rdh_theory}
\alias{rdh_theory}
\title{Calculates rugosity, fractal dimension or height range based the other two variables}
\usage{
rdh_theory(r, d, h, L, L0)
}
\arguments{
\item{r, d, h}{two of the three variables to calculate the third.}

\item{L}{extent.}

\item{L0}{resolution.}
}
\value{
A value corresponding one of the three variables not given to the function.
}
\description{
Calculates rugosity, fractal dimension or height range based the other two variables
}
\details{
This function uses a plane equation to calculate either rugosity,
fractal dimension or height range based on two of these variables.
}
\examples{
rdh_theory(r=3, h=1, L=2, L0=0.01)
rdh_theory(d=2.392472, h=1, L=2, L0=0.01)
rdh_theory(d=2.392472, r=3, L=2, L0=0.01)
}
\references{
Torres-Pulliza D, Dornelas M, Pizarro O, Bewley M, Blowes SA, Boutros N, Brambilla V, Chase TJ, Frank G, Friedman A, Hoogenboom MO, Williams S, Zawada KJA, Madin JS (2020) A geometric basis for surface habitat complexity and biodiversity. \emph{Nature Ecology & Evolution} 4:1495-1501. https://doi.org/10.1038/s41559-020-1281-8
}
