% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale}
\usage{
rescale(x, xmin, xmax, n)
}
\arguments{
\item{x}{vector}

\item{xmin}{minimum value}

\item{xmax}{maximum value}

\item{n}{multiplier}
}
\description{
Rescale
}
