% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{Rugosity}
\usage{
rg(
  data,
  x,
  y,
  L,
  L0,
  method = "dem",
  plot = FALSE,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{data}{Digital elevation model in raster format.}

\item{x}{Bottom-left of bounding box.}

\item{y}{Bottom-left of bounding box.}

\item{L}{Bounding box extent (i.e., side length).}

\item{L0}{Grain or resolution of calculation.}

\item{method}{"mesh" or "dem"}

\item{plot}{Add bounding box to plot.}
}
\value{

}
\description{
Rugosity
}
\examples{
rg(horseshoe, x=-470, y=1266, L=2, L0=0.5)

plot(horseshoe, asp=1)
rg(horseshoe, x=-470, y=1266, L=2, L0=0.5, plot=TRUE)

rg(mcap, L0 = 0.01, method = "mesh")

}
