% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_triangle.R
\name{sa_triangle}
\alias{sa_triangle}
\title{Calculate surface area of triangle}
\usage{
sa_triangle(XYZcoords)
}
\arguments{
\item{XYZcoords}{A data frame with XYZ coordinates of three points in following order: X1,X2,X3,Y1,Y2,Y3,Z1,Z2,Z3.}
}
\value{
The surface area of the triangle.
}
\description{
Calculates the surface area of a triangle based on a set of XYZCoords.
}
\examples{
sa_triangle(c(X1 = 1, X2 = 2, X3 = 3 , Y1 = 1, Y2 = 2, Y3 = 1, Z1 = 1, Z2 = 1, Z3 = 1))
}
