% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dem.R
\name{sample_dem}
\alias{sample_dem}
\title{Sample a random DEM with specified size from a larger DEM}
\usage{
sample_dem(data, L)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer.}

\item{L}{Size of square to cut out of DEM.}
}
\value{
Digital elevation model of class RasterLayer.
}
\description{
Sample a random DEM with specified size from a larger DEM
}
\examples{
dem1 <- sample_dem(horseshoe, L = 4)
raster::plot(dem1)
}
