% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_volume.R
\name{scale_volume}
\alias{scale_volume}
\title{Rescale mesh based on a fixed volume of 1}
\usage{
scale_volume(mesh)
}
\arguments{
\item{mesh}{A 3D mesh object}
}
\value{
A mesh with volume = 1
}
\description{
Rescale mesh based on a fixed volume of 1
}
\examples{
Rvcg::vcgVolume(mcap)
mcap_scaled <- scale_volume(mcap)
Rvcg::vcgVolume(mcap_scaled)
}
