% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_origin.R
\name{set_origin}
\alias{set_origin}
\title{Set the origin of a mesh}
\usage{
set_origin(mesh, reference = NULL)
}
\arguments{
\item{mesh}{A triangular mesh of class mesh3d.}

\item{reference}{Vector containing coordinates of the reference vertex.
If left empty, this will default to the minimum of x, y, and z.}
}
\value{
mesh3d object
}
\description{
Transforms XYZ coordinates relative to a chosen origin

Transforms coordinates so that the origin lies at the reference
vertex (defaults to the minimum of x, y, and z coordinates).
}
\examples{
mesh <- set_origin(mcap)
}
