% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dem.R
\name{sim_dem}
\alias{sim_dem}
\title{Simulates a fractal surface (terrain)}
\usage{
sim_dem(
  L,
  smoothness,
  H,
  R,
  plot = FALSE,
  prop = 0.1,
  n = 100,
  method = "area",
  parallel = FALSE
)
}
\arguments{
\item{L}{The extent.}

\item{smoothness}{A value between 0.0 and 1.0 (lower values
produce rougher DEM).}

\item{H}{Desired height range (optional).}

\item{R}{Desired rugosity value (optional).}

\item{n}{Number of iterations to try and reach desired R. Recommended to adapt R and H instead of increasing n if simulation fails.}

\item{method}{The method to be used for rugosity calculation in case R is given. Can be "hvar" or "area"}

\item{parallel}{Logical. Use parallel processing? Defaults to FALSE. Only relevant if method = "hvar".}
}
\value{
Digital elevation model of class RasterLayer.
}
\description{
Simulates z-values based on the Diamond-square algorithm.
Warning: this function gets slow for n > 128.
}
\examples{
library(raster)
dem <- sim_dem(L = 32, smoothness = 0.5)
plot(dem)
}
