% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_xyz.R
\name{sim_xyz}
\alias{sim_xyz}
\title{Simulate z values}
\usage{
sim_xyz(n, mu, phisq, sigmasq, dem = FALSE)
}
\arguments{
\item{n}{number of rows or cols}

\item{mu}{average z}

\item{phisq}{parameter}

\item{sigmasq}{parameter: maximum covariance}

\item{dem}{T or F}
}
\value{
a xyz dataframe or raster (if dem = T)
}
\description{
Simulate z values
}
\examples{
d <- sim_xyz(10, 6, 0.1, 2, dem = T)
plot(d)
}
