% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smv.R
\name{smv}
\alias{smv}
\title{Second moment volume}
\usage{
smv(mesh, axis = "z", scale = FALSE, origin = TRUE)
}
\arguments{
\item{mesh}{A 3D mesh object.}

\item{axis}{default = z.}

\item{scale}{To scale the object to have a volume = 1.}

\item{origin}{set the origin to the bottom left corner of bounding box}
}
\value{
second moment of volume
}
\description{
Calculates the 2nd moment of volume by multiplying the volume of each
triangle in the mesh by its distance from the origin (should be set to the attachment point
of the mesh). The sum of these values is the 2nd moment of volume.
To compare moments in terms of shape only, set scale = T.
Axis is z by default, meaning it will calculate the vertical second moment, but this can be changed if needed.
}
\examples{
smv(mcap)
smv(mcap, scale = T)
}
