% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terrain}
\alias{terrain}
\title{Simulates a fractal surface (terrain)}
\usage{
terrain(n, smoothness, z_extent = FALSE, dem = TRUE)
}
\arguments{
\item{n}{number of rows or cols}

\item{smoothness}{a value between 0.0 and 1.0 (lower values
produce rougher terrain)}

\item{z_extent}{scale z to match extent (default FALSE)}

\item{dem}{TRUE (default) or FALSE}
}
\value{
a matrix or raster (if dem = TRUE)
}
\description{
Simulates z-values based on the Diamond-square algorithm.
Warning: this function gets slow for n > 128.
}
\examples{
library(raster)
dem <- habtools::terrain(32, 0.5, z_extent=TRUE)
plot(dem)
}
