% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z}
\alias{z}
\title{Extract mean depth or elevation of a DEM}
\usage{
z(data)
}
\arguments{
\item{data}{A DEM in RasterLayer format.}
}
\value{
Value: mean depth or elevation of DEM.
}
\description{
Extract mean depth or elevation of a DEM
}
\examples{
z(horseshoe)
}
