% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count}
\alias{cell_count}
\title{Count filled cells}
\usage{
cell_count(pts, xmin, xmax, ymin, ymax, zmin, zmax, n)
}
\arguments{
\item{pts}{data}

\item{xmin}{minimum x-value}

\item{xmax}{maximum x-value}

\item{ymin}{minimum y-value}

\item{ymax}{maximum y-value}

\item{zmin}{minimum z-value}

\item{zmax}{maximum z-value}

\item{n}{multiplier}
}
\value{

}
\description{
Count filled cells
}
