% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count_2d}
\alias{cell_count_2d}
\title{Count filled cells in 2D}
\usage{
cell_count_2d(pts, xmin, xmax, ymin, ymax, n)
}
\arguments{
\item{pts}{Data frame with x and y coordinates}

\item{xmin}{Minimum x-value}

\item{xmax}{Maximum x-value}

\item{ymin}{Minimum y-value}

\item{ymax}{Maximum y-value}

\item{n}{Multiplier}
}
\value{
Number of filled cells
}
\description{
Count filled cells in 2D
}
