% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count.R
\name{cell_count_3d}
\alias{cell_count_3d}
\title{Count filled cells 3D}
\usage{
cell_count_3d(pts, xmin, xmax, ymin, ymax, zmin, zmax, n)
}
\arguments{
\item{pts}{Data frame with x, y, and z coordinates}

\item{xmin}{Minimum x-value}

\item{xmax}{Maximum x-value}

\item{ymin}{Minimum y-value}

\item{ymax}{Maximum y-value}

\item{zmin}{Minimum z-value}

\item{zmax}{Maximum z-value}

\item{n}{Multiplier}
}
\value{
Number of filled cells
}
\description{
A helper function for segment, box and cube counting fractal methods.
The function divide the array into \code{n} pieces and counts how many are
occupied.
}
