% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexity.R
\name{convexity}
\alias{convexity}
\title{Calculate convexity}
\usage{
convexity(mesh)
}
\arguments{
\item{mesh}{mesh3d object}
}
\value{
the convexity value
}
\description{
Calculate convexity
}
\details{
The ratio of the volume of the object  and the volume of the convex hull around the object
}
\examples{
convexity(mcap)
}
