% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_shape.R
\name{covar_shape}
\alias{covar_shape}
\title{Fit covariance-distance curve}
\usage{
covar_shape(data)
}
\arguments{
\item{data}{dataframe with, x, y, and z}
}
\value{
list including rhosq and range
}
\description{
Fit covariance-distance curve
}
\examples{
library(habtools)
raster <- raster::aggregate(horseshoe, 60)
data <- raster::rasterToPoints(raster)
covar_shape(data)

}
