% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csf.R
\name{csf}
\alias{csf}
\title{Calculates colony shape factor of a mesh}
\usage{
csf(mesh, L0, silent = FALSE)
}
\arguments{
\item{mesh}{mesh object}

\item{silent}{Print messages and warnings}
}
\value{
A value indicating colony shape factor
}
\description{
Calculates colony shape factor of a mesh
}
\examples{
csf(mcap)
}
