% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubes.R
\name{cubes}
\alias{cubes}
\title{Cube counting for 3D mesh}
\usage{
cubes(mesh, L0, plot = FALSE)
}
\arguments{
\item{mesh}{3D mesh with vertices}

\item{L0}{The grain or resolution (i.e., smallest cube size)}

\item{plot}{Plot number of filled cubes at different scales}
}
\value{
The fractal dimension value
}
\description{
Cube counting for 3D mesh
}
\examples{
cubes(mcap)

}
