% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_sample.R
\name{dem_sample}
\alias{dem_sample}
\title{Sample a random DEM with specified size from a larger DEM}
\usage{
dem_sample(data, L)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer.}

\item{L}{Size of square to cut out of DEM.}
}
\value{
Digital elevation model of class RasterLayer.
}
\description{
Sample a random DEM with specified size from a larger DEM
}
\examples{
dem <- dem_sample(horseshoe, L = 4)
raster::plot(dem)
}
