% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_mesh.R
\name{dem_to_mesh}
\alias{dem_to_mesh}
\title{DEM to 3D mesh}
\usage{
dem_to_mesh(dem)
}
\arguments{
\item{dem}{A digital elevation model in raster format}
}
\value{
A 3D points cloud for raster cell corners
}
\description{
This function is very slow and needs work.
}
\examples{
mesh <- dem_to_mesh(horseshoe)
# library(rgl)
plot3d(mesh)

}
