% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_to_points.R
\name{dem_to_points}
\alias{dem_to_points}
\title{Transform DEM to 3D pointcloud of raster corners.}
\usage{
dem_to_points(dem, bh = NULL, parallel = F)
}
\arguments{
\item{dem}{A digital elevation model in raster format}

\item{bh}{border height from lowest point of mesh}

\item{parallel}{TRUE or FALSE to define use of parallel computation}
}
\value{
A 3D point cloud for raster cell corners
}
\description{
Transform DEM to 3D pointcloud of raster corners.
}
