% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\title{Calculate extent of a 3D object}
\usage{
extent(data)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer or a triangular mesh of class mesh3d.}
}
\value{
A value, the extent of the mesh or DEM.
}
\description{
This function calculates the extent or largest length of the bounding box of a mesh or a DEM.
}
\note{
There are several extent function is other packages, including
the raster package. Therefore it is recommended to use the package
namespace, see examples below.
}
\examples{
habtools::extent(mcap)
habtools::extent(horseshoe)

}
