% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\title{Calculate extent of a 3D object}
\usage{
extent(data)
}
\arguments{
\item{data}{An object of class mesh3d}
}
\value{
A value, the extent of the mesh
}
\description{
Calculate extent of a 3D object
}
\details{
This function calculates the extent or largest length of the bounding box of a mesh
}
\examples{
habtools::extent(mcap)
habtools::extent(horseshoe)

}
