% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd}
\alias{fd}
\title{Fractal dimension}
\usage{
fd(data, method, x, y, L, lvec, ...)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer or a triangular mesh of class mesh3d.}

\item{method}{If data is a RasterLayer, possible methods are:"hvar", "area", "sd", and "cubes" (defaults to "hvar").
If data is a mesh3d, possible methods are "cubes" and "area" (defaults to "cubes").}

\item{x}{Bottom-left of bounding box. Only considered if data is a RasterLayer.}

\item{y}{Bottom-left of bounding box. Only considered if data is a RasterLayer.}

\item{L}{Bounding box extent (i.e., side length). Only considered if data is a RasterLayer.}

\item{lvec}{scales to use for calculation}

\item{...}{Arguments from other fd_ functions.}
}
\value{
A value for fractal dimension, typically between 2 and 3.
}
\description{
Fractal dimension
}
\details{
Calculates fractal dimension using the specified method. Note that methods are distinctly different and should not be mixed when comparing values for multiple objects.
The \code{cubes} method is not recommended if the height range is much smaller than the extent of a 3d object or DEM, which is typically the case for DEMs.
}
\examples{
library(habtools)
fd(horseshoe, method = "hvar", x = -470, y = 1266, L = 2, lvec = c(0.125, 0.25, 0.5, 1, 2))
fd(horseshoe, method = "area", x = -470, y = 1266, L = 2, lvec = c(0.125, 0.25, 0.5, 1, 2))
fd(mcap, method = "cubes", lvec = c(0.045, 0.09, 0.18, 0.5), plot = TRUE)
fd(mcap, method = "area", lvec = c(0.02, 0.04, 0.08, 0.16))

}
\seealso{
\code{\link[=fd_hvar]{fd_hvar()}}

\code{\link[=fd_area]{fd_area()}}

\code{\link[=fd_sd]{fd_sd()}}

\code{\link[=fd_cubes]{fd_cubes()}}
}
