% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{fd}
\alias{fd}
\title{Fractal Dimension}
\usage{
fd(data, method = "mean", plot = FALSE)
}
\arguments{
\item{data}{A \code{hvar} \code{data.frame}.}

\item{method}{One of \code{raw}, \code{mean} (default), \code{median} or \code{ends}}

\item{plot}{Diagnostic plot}
}
\value{
A value for fractal dimension, hopefully between 2 and 3.
}
\description{
Fractal Dimension
}
\details{
Calculates fractal dimension using the height variation method.
\code{data} can be any \code{data.frame} with columns labeled \code{L0} and \code{H0} for
grid cell length and height range of that cell, respectively.
A rule of thumb is that \code{L0} should range two orders of magnitude.
However, large ranges also
average-out fractal dimension of a surface that might have
phase transitions, and therefore a thorough exploration of height ranges is suggested using the \code{diagnostic} plot, which plots all the methods.
These methods will converge on purely fractal surfaces.
}
\examples{
data <- hvar(horseshoe, x=-470, y=1266, L=2, L0 = 0.5)
fd(data)
fd(data, method="raw")
fd(data, method="raw", plot=TRUE)

}
