% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_area.R
\name{fd_area}
\alias{fd_area}
\title{Calculate fractal dimension using the surface area method}
\usage{
fd_area(data, lvec, x, y, L, keep_data = FALSE, plot = FALSE)
}
\arguments{
\item{data}{dem of class "RasterLayer" or mesh class "mesh3d"}

\item{lvec}{vector of scales to include in calculation}

\item{x}{Bottom-left of bounding box.}

\item{y}{Bottom-left of bounding box.}

\item{L}{Bounding box extent (i.e., side length).}

\item{keep_data}{Keep area data. Default = FALSE.}

\item{plot}{Plot regression line and area data. Default = FALSE.}
}
\value{
Either a value or a list
}
\description{
Calculate fractal dimension using the surface area method
}
\examples{
fd_area(mcap, lvec = c(0.01, 0.02, 0.04, 0.08, 0.16))
fd_area(horseshoe, lvec = c(0.1, 0.2, 0.4, 0.8, 1.6))

}
