% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_cubes.R
\name{fd_cubes}
\alias{fd_cubes}
\title{Calculate fractal dimension using cube counting}
\usage{
fd_cubes(data, lvec = NULL, plot = FALSE, keep_data = FALSE, scale = FALSE)
}
\arguments{
\item{data}{An object of class RasterLayer or mesh3d.}

\item{lvec}{(optional) Scales to use for calculation (i.e. cube sizes).}

\item{plot}{Planar representation of cubes superimposed on 3D mesh for diagnosing \code{lvec}. Default = FALSE.}

\item{keep_data}{Logical. Keep calculation data? Default = TRUE.}

\item{scale}{Logical. Rescale height values to the extent. Only relevant for RasterLayer objects. (Defaults to FALSE).}
}
\value{
A \code{data.frame} of number of cubes (\code{n}) intersecting mesh points at different cube sizes (\code{L}) and a fractal dimension value. Note that cube size is the length of a side.
}
\description{
Calculate fractal dimension using cube counting
}
\details{
This function calculates fractal dimension using the cube counting method.
Based on lvec, cubes of different sizes are defined and the function counts mesh points that fall within each cube.
It is recommended that to specify the maximum value of \code{lvec} so that the largest box encapsulates the entire object.
The smallest scale included in \code{lvec} should not be smaller than the resolution of your object.
Most meshes are not perfectly fractal, and so use the \code{plot} parameter to look for scale transitions.
}
\examples{
fd_cubes(mcap, keep_data=TRUE, plot=TRUE)
fd_cubes(mcap, lvec = c(0.05, 0.1, 0.25, 0.5), plot=TRUE)

dem <- crop_dem(horseshoe, x0 = -469, y0 = 1267, L = 2, plot = TRUE)
fd_cubes(dem, plot=TRUE, keep_data=TRUE)
fd_cubes(dem, plot=TRUE, keep_data=TRUE, scale=TRUE)

}
\seealso{
\code{\link[=fd]{fd()}}
}
