% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\docType{data}
\name{horseshoe}
\alias{horseshoe}
\title{Horseshoe reef}
\format{
A 800 by 800 digital elevation model (of class RasterLayer).
\describe{
\item{Values}{depth}
\item{Resolution}{0.01 m}
\item{Extent}{8 m}
...
}
}
\usage{
horseshoe
}
\description{
A digital elevation model (DEM) of a reef patch in the Great Barrier Reef.
}
\examples{
raster::plot(habtools::horseshoe)
}
\keyword{datasets}
