% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hr}
\alias{hr}
\title{Height range}
\usage{
hr(data)
}
\arguments{
\item{data}{A digital elevation model (DEM) in a "RasterLayer" class or a "mesh3d" object.}
}
\value{
The difference between the lowest and highest point.
}
\description{
Calculates the distance between the lowest and highest point in a 3D object.
}
\examples{

library(raster)
plot(horseshoe)
hr(horseshoe)
hr(horseshoe)

# For a 3D mesh.
hr(mcap)

}
