% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hr}
\alias{hr}
\title{Height range}
\usage{
hr(data, method = "dem")
}
\arguments{
\item{data}{A 3D object.}

\item{method}{\code{dem} (default) is digital elevation model (DEM) in raster format; \code{mesh} is a 3D mesh model with vertices.}
}
\value{
The difference between the lowest and highest point in a 3D object.
}
\description{
Height range
}
\examples{
hr(horseshoe)

raster::plot(horseshoe)

# Within a bounding box
x <- -470
y <- 1266
L <- 2
rect(x, xy, x+L, y+L)

ext <- raster::extent(x, x+L, y, y+L)
hr(raster::crop(horseshoe, ext))

# For a 3D mesh.
hr(mcap, method="mesh")

}
