% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hr}
\alias{hr}
\title{Height range}
\usage{
hr(dem, x, y, L, plot = FALSE)
}
\arguments{
\item{dem}{A digital elevation model (DEM) in raster format.}

\item{x}{The lower-left x-coordinate of bounding box.}

\item{y}{The lower-left y-coordinate of bounding box.}

\item{L}{The bounding box size (i.e., side length).}

\item{plot}{Add bounding box to existing plot of \code{dem}.}
}
\value{
The difference between the lowest and highest point in the
bounding box.
}
\description{
Height range
}
\examples{
hr(horseshoe, -470, 1266, 2)

plot(horseshoe)
hr(horseshoe, -470, 1266, 2, plot=TRUE)

}
