% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hr}
\alias{hr}
\title{Calculate height range}
\usage{
hr(data)
}
\arguments{
\item{data}{A RasterLayer or mesh3d object.}
}
\value{
Value of height range.
}
\description{
Calculates the distance between the lowest and highest point in a 3D object.
}
\examples{
# for a DEM
hr(horseshoe)

# for a 3D mesh
hr(mcap)

}
