% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvar.R
\name{hvar}
\alias{hvar}
\title{Height Variation in cells at different scales}
\usage{
hvar(
  data,
  x,
  y,
  Lvec,
  L,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{data}{Digital elevation model of class RasterLayer.}

\item{x}{Bottom-left of bounding box.}

\item{y}{Bottom-left of bounding box.}

\item{Lvec}{scales to use for calculation}

\item{L}{Bounding box extent (i.e., side length).}

\item{parallel}{TRUE or FALSE. Use parallel processing? Note: parallel must be installed.}

\item{ncores}{number of cores to use when parallel = TRUE.}
}
\value{
A \code{data.frame} containing height ranges of cells at different scales.
}
\description{
Height Variation in cells at different scales
}
\details{
This function is used for calculating fractal dimension using the
height variation method.
}
\examples{

hvar(horseshoe, x = -470, y = 1266, L = 2, Lvec = c(0.125, 0.25, 0.5, 1))
}
