% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcap.R
\docType{data}
\name{mcap}
\alias{mcap}
\title{Montipora capitata}
\format{
mesh3d object with 5568 vertices, 10939 triangles.
}
\usage{
mcap
}
\description{
A laser scan of a coral colony.
}
\examples{
library(rgl)
plot3d(mcap)
}
\keyword{datasets}
