% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_2d.R
\name{mesh_to_2d}
\alias{mesh_to_2d}
\title{3D Mesh to 2D points}
\usage{
mesh_to_2d(mesh, L0 = NULL, plot = FALSE, silent = TRUE)
}
\arguments{
\item{mesh}{A mesh3d object.}

\item{L0}{(Optional) The desired DEM resolution in same units at the 3D mesh.}

\item{plot}{logical. Plot the output?}

\item{silent}{logical. Defaults to not showing warnings.}
}
\value{
A data frame.
}
\description{
\code{mesh_to_d2} turns a 3D Mesh file into an xy data frame.
}
\details{
The function rasterizes uses the vertices of the mesh file.
If resolution is not
given, it is calculated by finding the maximum nearest neighbor
of vertices projected
on the \code{xy} plane. \code{fill} is used when irregular 3D meshes
result in \code{NA} values in
raster cells. The default is to fill these cells with the
minimum, non-\code{NA} raster value.
}
\examples{
library(raster)
mcap_2d <- mesh_to_2d(mcap)

geometry::polyarea(mcap_2d$x, mcap_2d$y) # area
planar(mcap)

perimeter(mcap_2d) # perimeter
circularity(mcap_2d) # circularity
fd_boxes(mcap_2d) # fractal dimension

}
