% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_to_points.R
\name{mesh_to_points}
\alias{mesh_to_points}
\title{Transform mesh to 3D point cloud}
\usage{
mesh_to_points(mesh)
}
\arguments{
\item{mesh}{A triangular mesh of class mesh3d.}
}
\value{
A data frame with XYZ coordinates.
}
\description{
Transform mesh to 3D point cloud
}
