% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_find.R
\name{mid_find}
\alias{mid_find}
\title{Find midpoint of a DEM}
\usage{
mid_find(data)
}
\arguments{
\item{data}{a DEM in RasterLayer format}
}
\value{
A dataframe with x and y midpoints
}
\description{
Find midpoint of a DEM
}
\examples{
mid_find(horseshoe)
}
