% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{packing}
\alias{packing}
\title{Calculate packing}
\usage{
packing(mesh)
}
\arguments{
\item{mesh}{A triangular mesh of class mesh3d.}
}
\value{
Value of packing.
}
\description{
Calculate packing
}
\details{
The ratio of the surface area of the object and the surface area of the convex hull around the object.
}
\examples{
packing(mcap)
}
