% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter}
\alias{perimeter}
\title{The perimeter of a 2D shape}
\usage{
perimeter(data, keep_data = FALSE)
}
\arguments{
\item{data}{A data frame with the first two columns ordered x and y coordinates.}
}
\value{
The perimeter.
}
\description{
\code{mesh_to_d2} turns a 3D Mesh file into an xy data frame.
}
\details{
The function rasterizes uses the vertices of the mesh file.
}
\examples{
mcap_2d <- mesh_to_2d(mcap)
plot(mcap_2d)

perimeter(mcap_2d)

r <- 1 # radius
circ <- sim_circle(r=r) # simulate xy coordinates for a circle of radius 1
plot(circ, asp=1)
perimeter(circ)

2 * pi * r # Note xy resolution affects output

}
