% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planar.R
\name{planar}
\alias{planar}
\title{Calculates planar area of a mesh}
\usage{
planar(mesh, L0, silent = FALSE)
}
\arguments{
\item{mesh}{mesh object}

\item{L0}{Resolution of the planar area. Is set to the resolution of the mesh when left empty.}

\item{silent}{Print messages and warnings}
}
\value{
A value indicating planar area
}
\description{
Calculates planar area of a mesh
}
\examples{
planar(mcap)
}
