% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdh.R
\name{rdh}
\alias{rdh}
\title{Calculate Rugosity, fractal dimension and height for a DEM}
\usage{
rdh(
  data,
  x,
  y,
  L,
  L0,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{data}{A dem}

\item{x}{minimum x-value}

\item{y}{minimum y-value}

\item{L}{Extent}

\item{L0}{Minimum resolution}

\item{parallel}{use parallel processing}

\item{ncores}{number of cores to use if parallel = TRUE}
}
\value{
a dataframe with the three complexity measures
}
\description{
Calculate Rugosity, fractal dimension and height for a DEM
}
\examples{
rdh(horseshoe, x=-470, y=1266, L=2, L0=0.5)
}
