% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_area.R
\name{scale_area}
\alias{scale_area}
\title{Re-scale mesh based on a fixed area}
\usage{
scale_area(mesh, target_area = 1)
}
\arguments{
\item{mesh}{A triangular mesh of class mesh3d.}

\item{target_area}{The target area of the scaled 3D mesh. Defaults to 1.}
}
\value{
A mesh with area = target_area (1 as default).
}
\description{
Re-scale mesh based on a fixed area
}
\examples{
Rvcg::vcgArea(mcap)
mcap_scaled <- scale_area(mcap)
Rvcg::vcgArea(mcap_scaled)
}
