% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_terrain.R
\name{sim_terrain}
\alias{sim_terrain}
\title{Simulates a fractal surface (terrain)}
\usage{
sim_terrain(n, smoothness, z_extent = FALSE, dem = TRUE)
}
\arguments{
\item{n}{the extent (matrix size)}

\item{smoothness}{a value between 0.0 and 1.0 (lower values
produce rougher terrain)}

\item{z_extent}{logical. Should z be scaled to match extent, n?}

\item{dem}{logical. Should the matrix be turned into a DEM?}
}
\value{
a matrix or raster (if dem = TRUE)
}
\description{
Simulates z-values based on the Diamond-square algorithm.
Warning: this function gets slow for n > 128.
}
\examples{
library(raster)
dem <- habtools::terrain(32, 0.5, z_extent=TRUE)
plot(dem)
}
