% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\title{Second moment area}
\usage{
sma(mesh, axis = "z", scale = TRUE, origin = TRUE)
}
\arguments{
\item{mesh}{A mesh object}

\item{axis}{the axis along which to calculate the second moment of area. z is the default.}

\item{scale}{scale object so that total area = 1}

\item{origin}{set the origin to the bottom left corner of bounding box}
}
\value{
value
}
\description{
Calculates the 2nd moment of surface area by multiplying the surface area of each
triangle in the mesh by it's distance from the origin (should be set to the attachment point
of the mesh). The sum of these values is the 2nd moment of area.
}
\examples{
sma(mcap)
sma(mcap, scale = T)
}
