% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericity.R
\name{sphericity}
\alias{sphericity}
\title{Calculate sphericity}
\usage{
sphericity(mesh)
}
\arguments{
\item{mesh}{mesh 3d object}
}
\value{
value
}
\description{
Calculate sphericity
}
\details{
The ratio of the surface area of a sphere with the same volume as the object and the surface area of the object
}
\examples{
sphericity(mcap)
}
