% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericity.R
\name{sphericity}
\alias{sphericity}
\title{Calculate sphericity of a 3D object}
\usage{
sphericity(mesh)
}
\arguments{
\item{mesh}{A triangular mesh of class mesh3d.}
}
\value{
Sphericity value.
}
\description{
Calculates the ratio of the surface area of a sphere with the same volume as the object and the surface area of the object.
}
\examples{
sphericity(mcap)
}
\seealso{
\code{\link[=circularity]{circularity()}}
}
