% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dem.R
\name{split_dem}
\alias{split_dem}
\title{Split RasterLayer into smaller tiles}
\usage{
split_dem(data, size, parallel = FALSE, ncores = (parallel::detectCores() - 1))
}
\arguments{
\item{data}{RasterLayer}

\item{size}{size of tiles, relative to the unit of the RasterLayer}
}
\value{
List of RasterLayers
}
\description{
Split RasterLayer into smaller tiles
}
\examples{
L <- raster::res(horseshoe)[1] * nrow(horseshoe) # size of horseshoe = 8m
size <- 2 # size of target tiles
(L / size)^2 # number of target tiles = 16
dem_list <- split_dem(horseshoe, 2)
length(dem_list)

}
