% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_area.R
\name{surface_area}
\alias{surface_area}
\title{Calculate surface area}
\usage{
surface_area(data)
}
\arguments{
\item{data}{DEM in RasterLayer format, mesh3d object or data frame with xy coordinates.}
}
\value{
Surface area value.
}
\description{
Calculates surface area of a 3D or 2D object.
}
\examples{
surface_area(mcap)
surface_area(horseshoe)
surface_area(mesh_to_2d(mcap))

}
\references{
Jenness, J.S. Calculating Landscape Surface Area from Digital Elevation Models. Wildlife Society Bulletin, Vol. 32, No. 3 (Autumn, 2004), pp. 829-839n
}
