% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_area.R
\name{surface_area}
\alias{surface_area}
\title{Surface area}
\usage{
surface_area(data)
}
\arguments{
\item{data}{DEM in RasterLayer format, mesh3d object or data frame with xy coordinates}
}
\value{
surface area
}
\description{
Calculates surface area
}
\examples{
surface_area(mcap)
surface_area(horseshoe)
surface_area(mesh_to_2d(mcap))

}
